seal WriteState after populate stage to prevent (some) mutations
When writing a multi-threaded `Populator`, i.e. `runCycle` lambda,
a common enough error is to exit `populate()` before all tasks
have completed.

This change makes it so that `add(...)` and `getObjectMapper()`
will throw when called after the write state has been sealed
(which will happen after the populate stage completes).

Note: it *is* still possible to call `getStateEngine()` and get
into trouble. We cannot make it fail when sealed because we use
it internally.